+++ relay/channel/task/serviceinferencevideo/constants.go
@@ -1,0 +1,8 @@
+package serviceinferencevideo
+
+var ModelList = []string{
+	"doubao-seedance-2-0-260128",
+	"doubao-seedance-2-0-fast-260128",
+}
+
+var ChannelName = "service-inference.ai"